Console enabler works again
